requirements.txt
langchain
langchain_openai
langchain_community
python-dotenv
pypdf